refactor: consolidate globalStyle
