+++ apps/main/src/app/crvusd/client.tsx
@@ -8,7 +8,7 @@ import { OverlayProvider } from '@react-aria/overlays'
 import delay from 'lodash/delay'
 import { useCallback, useEffect, useState } from 'react'
 import { HashRouter } from 'react-router-dom'
-import GlobalStyle from '@/loan/globalStyle'
+import GlobalStyle from '@/globalStyle'
 import usePageVisibleInterval from '@ui-kit/hooks/usePageVisibleInterval'
 import Page from '@/loan/layout'
 import networks from '@/loan/networks'
