+++ apps/main/src/app/dao/client.tsx
@@ -14,7 +14,7 @@ import networks from '@/dao/networks'
 import useStore from '@/dao/store/useStore'
 import usePageVisibleInterval from '@ui-kit/hooks/usePageVisibleInterval'
 import Page from '@/dao/layout'
-import GlobalStyle from '@/dao/globalStyle'
+import GlobalStyle from '@/globalStyle'
 import { ChadCssProperties } from '@ui-kit/themes/typography'
 import { useUserProfileStore } from '@ui-kit/features/user-profile'
 import { useWallet } from '@ui-kit/features/connect-wallet'
