+++ apps/main/src/app/dex/client.tsx
@@ -9,7 +9,7 @@ import { useCallback, useEffect, useState } from 'react'
 import { HashRouter } from 'react-router-dom'
 import { persister, queryClient } from '@ui-kit/lib/api/query-client'
 import { ThemeProvider } from '@ui-kit/shared/ui/ThemeProvider'
-import GlobalStyle from '@/dex/globalStyle'
+import GlobalStyle from '@/globalStyle'
 import usePageVisibleInterval from '@ui-kit/hooks/usePageVisibleInterval'
 import Page from '@/dex/layout/default'
 import curvejsApi from '@/dex/lib/curvejs'
