+++ apps/main/src/app/lend/client.tsx
@@ -10,7 +10,7 @@ import { useCallback, useEffect, useState } from 'react'
 import { HashRouter } from 'react-router-dom'
 import { persister, queryClient } from '@ui-kit/lib/api/query-client'
 import { ThemeProvider } from '@ui-kit/shared/ui/ThemeProvider'
-import GlobalStyle from '@/lend/globalStyle'
+import GlobalStyle from '@/globalStyle'
 import Page from '@/lend/layout'
 import networks from '@/lend/networks'
 import { getPageWidthClassName } from '@/lend/store/createLayoutSlice'
