+++ apps/main/src/globalStyle.ts
